replace cobra with std lib (#710)
