+++ pkg/capabilities/cli/cmd/generate-types/go_workflow_builder.go.tmpl
