--- pkg/capabilities/cli/cmd/root_cmd.go
@@ -1,20 +1,0 @@
-package cmd
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/spf13/cobra"
-)
-
-var rootCmd = &cobra.Command{
-	Use:   "capability",
-	Short: "Tooling related to Baku capabilities",
-}
-
-func RootCmd() {
-	if err := rootCmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-}
