--- pkg/capabilities/cli/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/smartcontractkit/chainlink-common/pkg/capabilities/cli/cmd"
-)
-
-func main() {
-	cmd.RootCmd()
-}
